android jetpack pro/MyMovieList2/app/src/main/java/com/dicoding/helfani/mymovielist/ui/detail/DetailTvShowViewModel.kt
package com.dicoding.helfani.mymovielist.ui.detail

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import com.dicoding.helfani.mymovielist.data.MovieRepository
import com.dicoding.helfani.mymovielist.data.source.local.entity.TvShowEntity

class DetailTvShowViewModel(private val movieRepository: MovieRepository) : ViewModel() {

    private lateinit var tvShowId: String

    fun setSelectedTvShow(tvShowId: String) {
        this.tvShowId = tvShowId
    }

    fun getTvShow(): LiveData<TvShowEntity> = movieRepository.getTvShow(tvShowId)
}